+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 <div align="center"><h1>Changelog</h1></div>
 
-## [0.0.46](https://github.com/tanloong/neosca/releases/tag/0.0.45) (14 August 2023)
+## [0.0.47](https://github.com/tanloong/neosca/releases/tag/0.0.47) (18 August 2023)
+
+### Buf fixes
+
++ Re-calculate values of structures defined by value_source after '+' operation
++ Check empty key/value when loading custom config file
+
+## [0.0.46](https://github.com/tanloong/neosca/releases/tag/0.0.46) (14 August 2023)
 
 ### Bug fixes